# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "KB API",
        "description": "HTTP JSON API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Получение списка пользователей (admin)",
                "responses": {
                    "200": {
                        "description": "Список пользователей",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#components/schemas/UserResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неаутентифицирован"
                    },
                    "403": {
                        "description": "Неавторизован"
                    }
                },
                "security": {
                    "apiKey": {}
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Создание пользователя (admin)",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Созданный пользователь",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#components/schemas/UserResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неаутентифицирован"
                    },
                    "403": {
                        "description": "Неавторизован"
                    }
                },
                "security": {
                    "apiKey": {}
                }
            }
        },
        "/users/{user}": {
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Удаление пользователя (admin)",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "Идентификатор пользователя"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Удалено"
                    },
                    "401": {
                        "description": "Неаутентифицирован"
                    }
                },
                "security": {
                    "apiKey": {}
                }
            },
            "patch": {
                "tags": [
                    "Users"
                ],
                "summary": "Обновление пользователя (admin)",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "Идентификатор пользователя"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Обновленный пользователь",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#components/schemas/SectionWithFieldsResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неаутентифицирован"
                    }
                },
                "security": {
                    "apiKey": {}
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "UserCreateRequest": {
                "required": [
                    "login",
                    "email",
                    "name",
                    "password",
                    "role"
                ],
                "properties": {
                    "login": {
                        "type": "string",
                        "example": "ninja24"
                    },
                    "email": {
                        "type": "string",
                        "example": "some@domain.com"
                    },
                    "name": {
                        "type": "string",
                        "example": "Иванов Иван"
                    },
                    "password": {
                        "type": "string",
                        "example": "qwerty12345"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "user",
                            "moderator",
                            "admin"
                        ]
                    },
                    "photo": {
                        "type": "string",
                        "example": "some.com/123.jpg"
                    }
                },
                "type": "object"
            },
            "UserUpdateRequest": {
                "properties": {
                    "login": {
                        "type": "string",
                        "example": "ninja24"
                    },
                    "email": {
                        "type": "string",
                        "example": "some@domain.com"
                    },
                    "name": {
                        "type": "string",
                        "example": "Иванов Иван"
                    },
                    "password": {
                        "type": "string",
                        "example": "qwerty12345"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "user",
                            "moderator",
                            "admin"
                        ]
                    },
                    "photo": {
                        "type": "string",
                        "example": "some.com/123.jpg"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "apiKey": {
                "type": "apiKey",
                "description": "Api key",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "apiKey": []
        }
    ]
}